`LinearMixture` moved to `kesi._engine` module
