+++ lib/kesi/__init__.py
@@ -23,8 +23,6 @@
 ###############################################################################
 
 import numpy as np
-import operator
-import sys
 
 try:
     import pandas as pd
@@ -146,136 +144,6 @@ class KernelFieldApproximator(_KernelFieldApproximator):
             return tuple(map(self._ndarrays_to_tuples, point))
 
         return point
-
-
-class LinearMixture(object):
-    def __init__(self, components=[]):
-        self._components, self._weights = [], []
-
-        try:
-            for c, w in components:
-                if isinstance(c, LinearMixture):
-                    self._append_components_from_mixture(self._components,
-                                                         self._weights,
-                                                         self._components,
-                                                         c * w)
-                else:
-                    self._components.append(c)
-                    self._weights.append(w)
-
-        except TypeError:
-            self._components = (components,)
-            self._weights = (1,)
-
-        self._prepare_cache_for_dir()
-
-    def _prepare_cache_for_dir(self):
-        components = self._components
-        self._dir = ({attr for attr in dir(components[0])
-                      if all(hasattr(c, attr) for c in components[1:])}
-                     if components
-                     else ())
-
-    def __getattr__(self, name):
-        if name not in self._dir:
-            raise AttributeError
-
-        def wrapper(*args, **kwargs):
-            return sum(w * getattr(c, name)(*args, **kwargs)
-                       for w, c in zip(self._weights,
-                                       self._components))
-
-        return wrapper
-
-    def __dir__(self):
-        return list(self._dir)
-
-    def __add__(self, other):
-        return self._add(other)
-
-    def __radd__(self, other):
-        return self._add(other)
-
-    def _add(self, other):
-        if other == 0:
-            return self
-
-        components = list(self._components)
-        weights = list(self._weights)
-        self._append_components_from_mixture(components, weights,
-                                             self._components,
-                                             other)
-
-        return self.__class__(list(zip(components, weights)))
-
-    @staticmethod
-    def _append_components_from_mixture(components,
-                                        weights,
-                                        reference_components,
-                                        mixture):
-        for c, w in zip(mixture._components,
-                        mixture._weights):
-            try:
-                i = reference_components.index(c)
-            except ValueError:
-                weights.append(w)
-                components.append(c)
-            else:
-                weights[i] += w
-
-    def __neg__(self):
-        return -1 * self
-
-    def __sub__(self, other):
-        return self + -other
-
-    def __mul__(self, other):
-        return self._mul(other)
-
-    def __rmul__(self, other):
-        return self._mul(other)
-
-    def _mul(self, other):
-        if other == 1 or self == 0:
-            return self
-
-        return self.__class__([(c, w * other)
-                               for c, w
-                               in zip(self._components,
-                                      self._weights)])
-
-    def __truediv__(self, other):
-        return self._div(other)
-
-    def _div(self, other):
-        return 1. / other * self
-
-    def __eq__(self, other):
-        if self:
-            return self is other
-
-        return not other
-
-    def __ne__(self, other):
-        if self:
-            return self is not other
-
-        return bool(other)
-
-    def _bool(self):
-        return bool(self._components)
-
-    # Use API appropriate for current Python version
-    if sys.version_info.major > 2:
-        def __bool__(self):
-            return self._bool()
-
-    else:
-        def __nonzero__(self):
-            return self._bool()
-
-        def __div__(self, other):
-            return self._div(other)
 
 
 class FunctionalKernelFieldReconstructor(FunctionalFieldReconstructor):
